Make events use doubles

+++ src/entries/event.h
@@ -14,8 +14,8 @@ class Event : public Entry
     Q_PROPERTY(QString date READ getDate WRITE setDate)
     Q_PROPERTY(QString event READ getEvent WRITE setEvent)
     Q_PROPERTY(QString servicer READ getServicer WRITE setServicer)
-    Q_PROPERTY(float cost READ getCost WRITE setCost)
-    Q_PROPERTY(float increment READ getIncrement WRITE setIncrement)
+    Q_PROPERTY(double cost READ getCost WRITE setCost)
+    Q_PROPERTY(double increment READ getIncrement WRITE setIncrement)
     Q_PROPERTY(QString category READ getCategory WRITE setCategory)
     Q_PROPERTY(QString comment READ getComment WRITE setComment)
 
@@ -32,11 +32,11 @@ public:
     QString getServicer() const { return servicer; }
     void setServicer(QString servicer_) { servicer = servicer_; }
 
-    float getCost() const { return cost; }
-    void setCost(float cost_) { cost = cost_; }
+    double getCost() const { return cost; }
+    void setCost(double cost_) { cost = cost_; }
 
-    float getIncrement() const { return increment; }
-    void setIncrement(float increment_) { increment = increment_; }
+    double getIncrement() const { return increment; }
+    void setIncrement(double increment_) { increment = increment_; }
 
     QString getCategory() const { return category; }
     void setCategory(const QString &category_) { category = category_; }
@@ -48,8 +48,8 @@ private:
     QString date;
     QString event;
     QString servicer;
-    float cost;
-    float increment;
+    double cost;
+    double increment;
     QString category;
     QString comment;
 };
